wip: refacto to read from S3 instead of REST api
