+++ quotaclimat/data_processing/mediatree/api_import.py
@@ -17,6 +17,7 @@ from quotaclimat.data_processing.mediatree.detect_keywords import *
 from quotaclimat.data_processing.mediatree.channel_program import *
 from quotaclimat.data_processing.mediatree.stop_word.main import get_all_stop_word
 from quotaclimat.data_processing.mediatree.api_import_utils.db import get_last_date_and_number_of_delay_saved_in_keywords, KeywordLastStats
+from quotaclimat.data_processing.mediatree.s3.s3_utils import *
 from postgres.insert_data import save_to_pg
 from postgres.schemas.models import create_tables, connect_to_db, get_db_session
 from postgres.schemas.models import keywords_table
@@ -34,19 +35,6 @@ logging.getLogger('distributed.scheduler').setLevel(logging.ERROR)
 logging.getLogger('distributed').setLevel(logging.ERROR)
 logging.getLogger('worker').setLevel(logging.ERROR)
 sentry_init()
-
-#read whole file to a string
-password = get_password()
-AUTH_URL = get_auth_url()
-USER = get_user()
-KEYWORDS_URL = get_keywords_url()
-
-def refresh_token(token, date):
-    if is_it_tuesday(date): # refresh token every weekday for batch import
-        logging.info("refreshing api token every weekday in case it's expired")
-        return get_auth_token(password=password, user_name=USER)
-    else:
-        return token
 
 # reapply word detector logic to all saved keywords
 # use when word detection is changed
@@ -102,7 +90,7 @@ def get_channels():
 
     return channels
 
-def get_stop_words(session, validated_only=True, context_only=True):
+def get_stop_words(session, validated_only=True, context_only=True) -> list[Stop_Word]:
     logging.info("Getting Stop words...")
     try:
         stop_words = get_all_stop_word(session, validated_only=validated_only)
@@ -124,7 +112,7 @@ def get_stop_words(session, validated_only=True, context_only=True):
 
 def get_start_time_to_query_from(session)      :
     normal_delay_in_days = 1
-    lastSavedKeywordsDate = get_last_date_and_number_of_delay_saved_in_keywords(session)
+    lastSavedKeywordsDate: KeywordLastStats = get_last_date_and_number_of_delay_saved_in_keywords(session)
     logging.info(f"last saved date for keywords is {lastSavedKeywordsDate.last_day_saved}, with a delay of  \
                  {lastSavedKeywordsDate.number_of_previous_days_from_yesterday} days compared to yesterday")
     
@@ -146,36 +134,35 @@ def get_start_time_to_query_from(session)      :
         default_number_of_previous_days = lastSavedKeywordsDate.number_of_previous_days_from_yesterday
         return default_start_date, default_number_of_previous_days
 
-async def get_and_save_api_data(exit_event):
-    with sentry_sdk.start_transaction(op="task", name="get_and_save_api_data"):
+async def get_and_save_s3_data_to_pg(exit_event):
+    with sentry_sdk.start_transaction(op="task", name="get_and_save_s3_data_to_pg"):
         try:
             logging.warning(f"Available CPUS {os.cpu_count()} - MODIN_CPUS config : {os.environ.get('MODIN_CPUS', 3)}")
 
             conn = connect_to_db()
             session = get_db_session(conn)
-            token=get_auth_token(password=password, user_name=USER)
-            type_sub = 's2t'
 
             (start_date, number_of_previous_days) = get_start_time_to_query_from(session)
-            # TODO : Getting date range from 2024-09-29 00:00:00 - 2024-09-15 00:00:00
-            # date should be added
             (start_date_to_query, end_date) = get_start_end_date_env_variable_with_default(start_date, minus_days=number_of_previous_days)
             df_programs = get_programs()
             channels = get_channels()
             
             stop_words = get_stop_words(session, validated_only=True)
             
-            
             day_range = get_date_range(start_date_to_query, end_date, number_of_previous_days)
             logging.info(f"Number of days to query : {len(day_range)} - day_range : {day_range}")
             for day in day_range:
-                # TODO should query S3 directly
-                token = refresh_token(token, day)
-                
                 for channel in channels:
                     try:
                         logging.info("Querying day %s for channel %s" % (day, channel))
                         programs_for_this_day = get_programs_for_this_day(day.tz_localize("Europe/Paris"), channel, df_programs)
+
+                        # get s3 path from year,month,day and media
+                        s3_path: str = get_bucket_key_folder(date=day, channel=channel)
+                        df_channel_for_a_day = read_folder_from_s3(s3_path)
+
+                        # should map row to update in case program name
+                        # df_channel_for_a_day.apply ()....
 
                         for program in programs_for_this_day.itertuples(index=False):
                             start_epoch = program.start
@@ -184,10 +171,7 @@ async def get_and_save_api_data(exit_event):
                             channel_program_type = str(program.program_type)
                             logging.info(f"Querying API for {channel} - {channel_program} - {channel_program_type} - {start_epoch} - {end_epoch}")
 
-                            # TODO should query S3
-                            # df = extract_s3_sub(token, channel, type_sub, start_epoch,end_epoch, channel_program, \
-                            #                      channel_program_type, stop_words=stop_words)
-                            df = extract_api_sub(token, channel, type_sub, start_epoch,end_epoch, channel_program, \
+                            df = extract_api_sub(channel, start_epoch, end_epoch, channel_program, \
                                                  channel_program_type, stop_words=stop_words)
                             if(df is not None):
                                 logging.debug(f"Memory df {df.memory_usage()}")
@@ -201,42 +185,10 @@ async def get_and_save_api_data(exit_event):
                         continue
             exit_event.set()
         except Exception as err:
-            logging.fatal("get_and_save_api_data (%s) %s" % (type(err).__name__, err))
+            logging.fatal("get_and_save_s3_data (%s) %s" % (type(err).__name__, err))
             ray.shutdown()
             sys.exit(1)
 
-# "Randomly wait up to 2^x * 1 seconds between each retry until the range reaches 60 seconds, then randomly up to 60 seconds afterwards"
-# @see https://github.com/jd/tenacity/tree/main
-@retry(wait=wait_random_exponential(multiplier=1, max=60),stop=stop_after_attempt(7))
-def get_auth_token(password=password, user_name=USER):
-    logging.info(f"Getting a token for user {user_name}")
-    try:
-        post_arguments = {
-            'grant_type': 'password'
-            , 'username': user_name
-            , 'password': password
-        }
-        response = requests.post(
-            AUTH_URL, 
-            data=post_arguments
-        )
-        output = response.json()
-        token = output['data']['access_token']
-        return token 
-    except Exception as err:
-        logging.error("Could not get token %s:(%s) %s" % (type(err).__name__, err))
-
-# see : https://keywords.mediatree.fr/docs/#api-Subtitle-SubtitleList
-def get_param_api(token, type_sub, start_epoch, channel, end_epoch):
-
-    return {
-        "channel": channel,
-        "token": token,
-        "start_gte": int(start_epoch) - EPOCH__5MIN_MARGIN,
-        "start_lte": int(end_epoch) + EPOCH__5MIN_MARGIN,
-        "type": type_sub,
-        "size": "1000" #  range 1-1000
-    }
 
 # "Randomly wait up to 2^x * 1 seconds between each retry until the range reaches 60 seconds, then randomly up to 60 seconds afterwards"
 # @see https://github.com/jd/tenacity/tree/main
@@ -254,16 +206,6 @@ def get_post_request(media_tree_token, type_sub, start_epoch, channel, end_epoch
         return parse_raw_json(response)
     except Exception as err:
         logging.error("Retry - Could not query API :(%s) %s" % (type(err).__name__, err))
-        raise Exception
-
-@retry(wait=wait_random_exponential(multiplier=1, max=60),stop=stop_after_attempt(7))
-def get_df_api(media_tree_token, type_sub, start_epoch, channel, end_epoch, channel_program, channel_program_type):
-    try:
-        response_sub = get_post_request(media_tree_token, type_sub, start_epoch, channel, end_epoch)
-
-        return parse_reponse_subtitle(response_sub, channel, channel_program, channel_program_type)
-    except Exception as err:
-        logging.error("Retry - get_df_api:(%s) %s" % (type(err).__name__, err))
         raise Exception
 
 # Data extraction function definition
@@ -305,56 +247,10 @@ def parse_total_results(response_sub) -> int :
 def parse_number_pages(response_sub) -> int :
     return int(response_sub.get('number_pages'))
 
-def parse_reponse_subtitle(response_sub, channel = None, channel_program = "", channel_program_type = "") -> Optional[pd.DataFrame]:
-    with sentry_sdk.start_transaction(op="task", name="parse_reponse_subtitle"):
-        total_results = parse_total_results(response_sub)
-        logging.getLogger("modin.logging.default").setLevel(logging.WARNING)
-        if(total_results > 0):
-            logging.info(f"{total_results} 'total_results' field")
-           
-            # To avoid  UserWarning: json_normalize is not currently supported by PandasOnRay, defaulting to pandas implementation.
-            flattened_data = response_sub.get("data", [])
-            new_df : pd.DataFrame = pd.DataFrame(flattened_data)
-            new_df["channel.name"] = new_df["channel"].apply(lambda x: x["name"])
-            new_df["channel.title"] = new_df["channel"].apply(lambda x: x["title"])
-            new_df["channel.radio"] = new_df["channel"].apply(lambda x: x["radio"])
-            new_df.drop("channel", axis=1, inplace=True)
-
-            logging.debug("Schema from API before formatting :\n%s", new_df.dtypes)
-            pd.set_option('display.max_columns', None)
-            logging.debug("setting timestamp")
-            new_df['timestamp'] = new_df.apply(lambda x: pd.to_datetime(x['start'], unit='s', utc=True), axis=1)
-            logging.debug("timestamp was set")
-
-            logging.debug("droping start column")
-            new_df.drop('start', axis=1, inplace=True)
-            logging.debug("renaming columns")
-            new_df.rename(columns={'channel.name':'channel_name', 
-                                   'channel.title':'channel_title',
-                                   'channel.radio': 'channel_radio',
-                                    'timestamp':'start'
-                                  },
-                        inplace=True
-            )
-
-            logging.debug("setting channel_title")
-            new_df['channel_title'] = new_df.apply(lambda x: get_channel_title_for_name(x['channel_name']), axis=1)
-
-            logging.debug(f"setting program {channel_program}")
-            # weird error if not using this way: (ValueError) format number 1 of "20h30 le samedi" is not recognized
-            new_df['channel_program'] = new_df.apply(lambda x: channel_program, axis=1)
-            new_df['channel_program_type'] = new_df.apply(lambda x: channel_program_type, axis=1)
-            logging.debug("programs were set")
-           
-            return new_df
-        else:
-            logging.warning("No result (total_results = 0) for this channel")
-            return None
-
 async def main():
     with monitor(monitor_slug='mediatree'): #https://docs.sentry.io/platforms/python/crons/
         try:
-            logging.info("Start api mediatree import")
+            logging.info("Start api mediatree import from S3")
             create_tables()
 
             event_finish = asyncio.Event()
@@ -372,7 +268,7 @@ async def main():
             if(os.environ.get("UPDATE") == "true"):
                 asyncio.create_task(update_pg_data(event_finish))
             else:
-                asyncio.create_task(get_and_save_api_data(event_finish))
+                asyncio.create_task(get_and_save_s3_data_to_pg(event_finish))
 
             # Wait for both tasks to complete
             await event_finish.wait()
